Refactor ImageSlider to Left/Right arrow for easier positioning
