+++ client/src/components/overview/DefaultImages/DefaultImages.css
@@ -1,11 +1,14 @@
 .gallery {
-  width: 450px;
-  height: 600px;
+  width: 50%;
+  height: 70%;
+  display: flex;
+  justify-content: space-between;
 }
 
 .images {
-  width: 500px;
-  height: 650px;
+  object-fit: contain;
+  max-width: 100%;
+  max-height: 100%;
   cursor: zoom-in;
 }
 
@@ -16,4 +19,7 @@
 .currImage {
   width: 500px;
   height: 650px;
+  display: flex;
+  justify-content: center;
+  align-content: center;
 }
