+++ client/src/components/overview/DefaultImages/DefaultImages.jsx
@@ -1,7 +1,8 @@
 import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 import Styles from './DefaultImages.css';
-import ImageSlider from './ImageGallery/ImageGallery.jsx';
+import LeftArrow from './Arrows/LeftArrow.jsx';
+import RightArrow from './Arrows/RightArrow.jsx';
 import ProductThumbnails from './ProductThumbnails/ProductThumbnails.jsx';
 import Modal from './Modal/Modal.jsx';
 
@@ -12,13 +13,17 @@ const DefaultImages = ({ photos }) => {
   if (photos) {
     return (
       <div className={Styles.gallery}>
+
+        <LeftArrow currIndex={index} updateIndex={setIndex} length={photos.length} />
+
         <div className={Styles.currImage} onClick={() => { setModal(true); }}>
           <img className={Styles.images} src={photos[index].url} alt="defaultImage" />
         </div>
 
+        <RightArrow currIndex={index} updateIndex={setIndex} length={photos.length} />
+
         <Modal handleClose={setModal} show={modal} currIndex={index} updateIndex={setIndex} photos={photos}/>
 
-        <ImageSlider currIndex={index} updateIndex={setIndex} length={photos.length} />
         <ProductThumbnails productThumbnails={photos} currIndex={index} updateIndex={setIndex} />
       </div>
     );
