--- client/src/components/overview/DefaultImages/ImageGallery/ImageGallery.css
@@ -1,41 +1,0 @@
-.slider {
-  min-width: 600px;
-  top: 50%;
-  left: 8%;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  position: absolute;
-}
-
-.arrow {
-  min-width: 50px;
-  min-height: 50px;
-}
-
-.LeftArrow {
-  width: 50px;
-  height: 50px;
-  top: 50%;
-  z-index: 2;
-  cursor: pointer;
-  user-select: none;
-}
-
-.RightArrow {
-  width: 50px;
-  height: 50px;
-  top: 50%;
-  z-index: 2;
-  cursor: pointer;
-  user-select: none;
-}
-
-.CurrentPic {
-  opacity: 1;
-  transition-duration: 1s;
-}
-
-.Hidden {
-  visibility: hidden
-}
