+++ client/src/components/overview/DefaultImages/Modal/Modal.css
@@ -7,9 +7,10 @@
   background: rgba(0, 0, 0, 0.6);
 }
 
-.images {
+.image {
   width: 70%;
   height: 85%;
+  cursor: crosshair;
 }
 
 .currImage {
